fix: api 도메인 분리
