+++ src/feature/attendance/api/index.ts
@@ -1,2 +1,3 @@
 export * from "./hooks";
 export * from "./types";
+export type { AttendanceStatus } from "@/feature/session/api";
