+++ src/feature/attendance/api/types.ts
@@ -1,51 +1,4 @@
-/**
- * 부트캠프 정보
- */
-export interface Bootcamp {
-  id: number;
-  name: string;
-  organizer: string;
-  isKdt: boolean;
-  startDate: string;
-  endDate: string;
-}
-
-/**
- * 부트캠프 목록 응답
- */
-export interface BootcampSummaryResponse {
-  totalCount: number;
-  bootcamps: Bootcamp[];
-}
-
-/**
- * 세션 정보
- */
-export interface Session {
-  id: number;
-  bootcampId: number;
-  classDate: string;
-  unitNo: number;
-  periodStartDate: string;
-  periodEndDate: string;
-  attendance?: {
-    status: AttendanceStatus;
-    userId: number;
-  };
-}
-
-/**
- * 출결 상태
- */
-export type AttendanceStatus =
-  | "present"
-  | "absent"
-  | "late"
-  | "left_early"
-  | "annual"
-  | "leave"
-  | "no_session"
-  | "no_attendance";
+import type { AttendanceStatus } from "@/feature/session/api";
 
 /**
  * Calendar 컴포넌트에서 사용하는 출결 상태
