+++ src/feature/attendance/index.ts
@@ -1,1 +1,2 @@
 export * from "./api";
+export type { AttendanceStatus } from "@/feature/session/api";
